Add more JSON routes and be more RESTful
All the core now routes take an optional ".<fmt>" suffix, which
defaults to "*".  The two allowed formats (other than for static files
and cover images) are `application/json` and `text/html`: if `fmt` is
"json" or "html", and the `Accept` header matches, then the response
will be returned in that format.

The core routes are now:

    GET    /search              -> do search
    GET    /book/<id>           -> get book details (only json)
    PUT    /book/<id>           -> update book
    DELETE /book/<id>           -> delete book
    GET    /add                 -> render add form (only html)
    POST   /add                 -> create book
    GET    /book/<id>/edit      -> render edit form (only html)
    POST   /book/<id>/edit      -> alias for PUT /book/<id>
    GET    /book/<id>/delete    -> render delete form (only html)
    POST   /book/<id>/delete    -> alias for DELETE /book/<id>
    GET    /book/<id>/cover     -> get cover image (multiple formats)
    GET    /static/<file>       -> get static file (multiple formats)

The following redirects exist for legacy purposes:

    GET    /                    -> redirect to /search
    GET    /list                -> redirect to /search

As these only exist to redirect any existing links, they do not have
the optional ".<fmt>" suffix.