Add support for reference books
aka remove the yes/no has-been-read dichotomy!

Having a growing number of unread books makes me feel bad, so this
commit (1) adds a third state: "not-applicable", which is surfaced in
the UI as "is a reference book" more-or-less, and (2) removes the line
about what proportion of books have been read.

An unexpected upside is that this allows simplifying some code, as I can
delete some aggregations and template variables that I don't need any
more.

As this changes the ES schema, for search to work the index will have to
be dumped and restored, but that's an easy operation now as prior
commits in this PR make it possible to parse old-format data.